2006-08-15  Theppitak Karoonboonyanan  <thep@linux.thai.net>
	* datrie/Makefile.am, +datrie/tail.c, datrie/tail.h: Added tail.c for
	trie suffix implementation. Adjusted some API to not require size_t.

	* datrile/fileutils.h: Added more functions required by tail.c.

	* datrie/trie.c (trie_branch_in_tail): Added check for null suffix.

+++ datrie/tail.h
@@ -135,11 +135,11 @@ Bool     tail_set_data (Tail *t, TrieIndex index, TrieData data);
  * @a *suffix_idx is updated to the position after the last successful walk,
  * and the function returns the total number of character succesfully walked.
  */
-size_t   tail_walk_str  (Tail            *t,
+int      tail_walk_str  (Tail            *t,
                          TrieIndex        s,
                          short           *suffix_idx,
                          const TrieChar  *str,
-                         size_t           len);
+                         int              len);
 
 /**
  * @brief Walk in tail with a character
